+++ src/routes/__init__.py
@@ -1,0 +1,1 @@
+from .users import users_api
